Populate Redux store using sagas
This patch implements a mechanism to retrieve the fatality data using
redux-sagas. At this time, mock data are used, but it will be fairly
easy to replace it with actual API calls.

By using mock data retrieved via saga, we are able to simulate API
calls, for instance by adding some delay before returning the data. This
allows us to keep working on the next bricks needed for the project
(graphs, maps, etc.)