Add date filtering
Adds logic to the saga to filter the dates it retrieves. As we won't have
any API at first, we will simply retrieve the full data set when the
page loads for the first time (it will be cached by the service worker
after that), and filter out the information we don't want to display.

Fixes scrapd/scrapdviz#27