Prepare the preview version (#44)
This version is functional, but does not import the `antd` CSS for some
reason. In spite of that the data is retrieved and the graphs are
displayed correctly. As a result we can start using it as a baseline for
the deployments.

The GitHub templates were updated to use the correct URLs, pointing to
the `ScrAPD` organization.