Create a home page (#149)
* Add developers page linking to source code, docs, and open austin

* Complete rework of the page - definitey needs another overhaul but the base structure is there

* Fixing lint issues

* Reorganize

* Reorganize

* Add HomeSection component (#150)

Adds a component to create the sections defined on the home page.

* Add Hero Component (#151)

* Add Hero Component

* Update carImage variable name to heroImage

Updates variable to make it less specific.

* Delete unused JS objects

Deletes JS objects that were in Jeeth's commit that are no longer being used. This allowed the project to pass CI tests

* Scrapd hero css in js (#153)

* Add Hero Component

* Delete unused JS objects

Deletes JS objects that were in Jeeth's commit that are no longer being used. This allowed the project to pass CI tests

* Refactor CSS in JS

Refactors Hero Component.

* Adds properly named photos to home page (#157)

* Adds photos to homepage, assets folder
* Renames assets from last commit
* Updates component to pass CI test

* Update home page content (#158)

Updates the content of the home page to provide more relevant
information to the users.